+++ app/settings.py
@@ -11,10 +11,10 @@ LOGGING_LEVEL = os.getenv("LOGGING_LEVEL", "DEBUG")
 HEALTHCHECK_DELAY = os.getenv("HEALTHCHECK_DELAY", 300)
 
 RABBIT_URL = 'amqp://{user}:{password}@{hostname}:{port}/{vhost}'.format(
-    hostname=os.getenv('SEFT_RABBITMQ_HOST', 'localhost'),
+    hostname=os.getenv('SEFT_RABBITMQ_HOST', '127.0.0.1'),
     port=os.getenv('SEFT_RABBITMQ_PORT', 5672),
-    user=os.getenv('SEFT_RABBITMQ_DEFAULT_USER', 'rabbit'),
-    password=os.getenv('SEFT_RABBITMQ_DEFAULT_PASS', 'rabbit'),
+    user=os.getenv('SEFT_RABBITMQ_DEFAULT_USER', 'guest'),
+    password=os.getenv('SEFT_RABBITMQ_DEFAULT_PASS', 'guest'),
     vhost=os.getenv('SEFT_RABBITMQ_DEFAULT_VHOST', '%2f')
 )
 
@@ -32,14 +32,10 @@ RABBIT_QUEUE = os.getenv("SEFT_RABBIT_CONSUMER_QUEUE", "Seft.Responses")
 RABBIT_EXCHANGE = os.getenv('SEFT_RABBITMQ_EXCHANGE', 'message')
 RABBIT_QUARANTINE_QUEUE = os.getenv("SEFT_RABBIT_CONSUMER_QUARANTINE_QUEUE", "Seft.Responses.Quarantine")
 
-# ras keys
-RAS_SEFT_CONSUMER_PUBLIC_KEY = get_key(os.getenv('RAS_SEFT_CONSUMER_PUBLIC_KEY', "./test_keys/sdc-seft-signing-ras-public-key.pem"))
-
-# sdx keys
-SDX_SEFT_CONSUMER_PRIVATE_KEY = get_key(os.getenv('SDX_SEFT_CONSUMER_PRIVATE_KEY', "./test_keys/sdc-seft-encryption-sdx-private-key.pem"))
-
 FTP_HOST = os.getenv('SEFT_FTP_HOST', 'localhost')
 FTP_PORT = int(os.getenv('SEFT_FTP_PORT', '2021'))
 FTP_USER = os.getenv('SEFT_FTP_USER', 'ons')
 FTP_PASS = os.getenv('SEFT_FTP_PASS', 'ons')
 FTP_FOLDER = os.getenv('SEFT_CONSUMER_FTP_FOLDER', '.')
+
+SDX_KEYS_FILE = os.getenv('SDX_KEYS_FILE', 'keys.yml')