+++ app/tests/test_healthchecks.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch
-from app.main import HealthCheck
+from app.main import SetHealth
 from app.sdxftp import SDXFTP
 from ftplib import Error
 import pika
@@ -19,12 +19,15 @@ class TestHealthCheck(unittest.TestCase):
 
     @unittest.skipIf(connect_to_rabbit() == "fail", "no running rabbitmq")
     def test_rabbit_mq_health_check(self):
-        self.assertEqual(HealthCheck.rabbit_health(), "ok")
+        set_health = SetHealth()
+        self.assertEqual(set_health.rabbit_status(), True)
 
     @patch.object(SDXFTP, '_connect')
     def test_ftp_healthcheck_success(self, mock_conn):
-        self.assertEqual(HealthCheck.ftp_health(), "ok")
+        set_health = SetHealth()
+        self.assertEqual(set_health.ftp_status(), True)
 
     @patch.object(SDXFTP, '_connect', side_effect=Error)
     def test_on_ftp_healthcheck_error_ftp_health_fails(self, mock_conn):
-        self.assertEqual(HealthCheck.ftp_health(), "failed")
+        set_health = SetHealth()
+        self.assertEqual(set_health.ftp_status(), False)