+++ CHANGELOG.md
@@ -2,3 +2,5 @@
   - Initial version of the SEFT publisher service
   - Namespace config
   - FTP port must be an int
+  - Updated to use SDC crypto library
+  - Add healthcheck endpoint